Build Rust executables
Rename 'app' to 'app_c', and add 'app_rust' to demo the new
`add_rust_executable()` CMake function.

In 'app_rust', I also demo using a relative path to bring in Rust crates
from the same project.

I added a 'build.rs' file for both running unit tests for the Rust
library, to automatically build with cbindgen and bindgen, and also a
'build.rs' for linking 'app_rust' with any C library dependencies, and
running 'bindgen' for that as well.

In the 'gen_uuid' module, I now export both a native rust API as well as
an `extern "C"` API so it can be used from both C and from Rust.