FindRust upgrades, integrate c/bindgen w. build.rs
`FindRust.cmake` no longer attempts to guess at the Rust platform target
triple by evaluating the OS, Architecture, and pointer-size of the target except
on Windows (for cross-compiling to x86) and on macOS (for universal binary builds).
Instead, it uses Rust's host target triple obtained with the
command `rustc -vV`.
If you want to cross-compile, you can set the `RUST_COMPILER_TARGET`
variable.

`FindRust.cmake` now sets the Rust build type to Release for CMake's
MinSizeRel and RelWithDebInfo builds, using the `-g` RUSTFLAG for the
latter so that debug symbols are provided.

Renamed a bunch of things to make it more intuitive. Namely the demo
library is now called "demo" instead of "CMakeRust", and the C and Rust
sources are now in separate directories such that the Rust stuff follows
the traditional Rust project layout.

Demo now integrates the bindgen and cbindgen processes into a `build.rs`
file that automatically runs at build time and does not depend on having
bindgen or cbindgen pre-installed. Generating C bindings with cbindgen
works everywhere, but generating the Rust bindings from the C code
appears to require libclang and may not be present on all systems - so
we only run it when MAINTAINER_MODE is ON.

Demo now installs the Rust-built static library.

The demo app is linked with the static libs so we don't have to make
a temporary install on Windows so the app can load the DLL.
Similarly, we link with the static libs for Rust unit test executable, so we
don't need a temporary install to run the test executable.

Also add clang-format spec for convenience in maintaining the C bits.

For Github Actions: switched to using CMake manually.