Specify source AND binary dirs in `add_rust_*` fun. Breaks API!
Changed the `add_rust_*` functions to have SOURCE_DIRECTORY and
BINARY_DIRECTORY parameters instead of WORKING_DIRECTORY.
This allows you to build and run the Rust unit tests from the same
BINARY_DIRECTORy as you do for the Rust library build. That means it
won't recompile all of the dependencies when building the unit test
executable.

A side-effect of this change is that the CMakeLists.txt file that uses
`add_rust_library()` must be in the same directory as the Rust module.
I.e. you should co-locate the CMakeLists.txt with the Cargo.toml file.
You can no longer have a parent directory's CMakeFiles.txt add a
library for a subdirectory.
The reason is that CMake will set the INTERFACE_INCLUDE_DIRECTORIES
property for your library's CMake target use both the SOURCE_DIRECTORY
and BINARY_DIRECTORY. This ensures that generated headers will be placed
in the include path for any downstream CMake targets.
CMake requires both directories to exist. When CMake does the build, it
will create subdirectories in the build path for each CMakeLists.txt
file. Co-locating CMakeLists.txt with Cargo.toml ensures that the
BUILD_DIRECTORY is automatically created by CMake.

Important: This commit breaks the API for the FindRust.cmake module!