Fix build issue when MAINTAINER_MODE=OFF / not set
It appears that when MAINTAINER_MODE=ON, the change in that environment
variable for the demorust ( lib/rust ) module causes the build.rs script
to be re-run.  I think that's because the MAINTAINER_MODE variable may
not be passed to `cargo test` from `ctest`, while it is passed to `cargo
build` from `cmake`.  However, when MAINTAINER_MODE=OFF there is no
change in the environment variables passed and so `build.rs` is not
re-run for `cargo test`.  As a consequence, it never links the test
executable with the required librar(ies).

The primary change in this commit to fix that is to move:
```rust
println!("cargo:rerun-if-env-changed=LIBDEMO");
```
out of the "test" match case so it applies in general.
That seems to fix it. I'm ... not 100% sure why, but it works for me!